+++ examples/data-sources/skycloak_extensions/data-source.tf
@@ -1,0 +1,5 @@
+data "skycloak_extensions" "all" {}
+
+output "marketplace_extensions" {
+  value = [for e in data.skycloak_extensions.all.extensions : e.name if e.source == "marketplace"]
+}